+++ ts/vm-asyncify.ts
